settings/django/requirements.txt
Django>=3.2.3,<3.3
gunicorn==20.1.0
django-smart-selects==1.5.9
djangorestframework==3.13.1
redis==4.5.1
gunicorn==20.1.0
django-model-utils==4.3.1
django-notifications-hq==1.8.2
djangorestframework-simplejwt==5.3.0
django-nested-admin==4.0.2
psycopg2==2.9.5
